+++ src/shared/constants.ts
@@ -4,7 +4,7 @@
 
 // Global application constants
 
-export const APP_VERSION = '0.26.1'
+export const APP_VERSION = '0.27.0'
 export const APP_NAME = 'SPERT Forecaster'
 export const APP_NAME_WITH_TM = 'SPERT® Forecaster'
 export const APP_FULL_NAME = 'SPERT® Suite'